spacecraft wip stable v4

+++ app/src/main/java/com/ardovic/farkle/dice/game/Coordinate.kt
@@ -1,3 +1,3 @@
 package com.ardovic.farkle.dice.game
 
-class Coordinate(var x: Float, var y: Float)
+class Coordinate(var x: Int, var y: Int)
